refactor/prepare-for-klap-protocol
