+++ .vscode/settings.json
@@ -23,5 +23,5 @@
       }
     }
   },
-  "cSpell.words": ["Tapo"]
+  "cSpell.words": ["KLAP", "Tapo"]
 }
