+++ src/api/@types/Protocol.ts
@@ -1,0 +1,6 @@
+enum Protocol {
+  Legacy,
+  KLAP
+}
+
+export default Protocol;
